check access add to function and btn

+++ src/component/person/PersonManage/PersonManage.tsx
@@ -18,6 +18,7 @@ import 'moment/locale/fa';
 import 'moment/locale/ar';
 import moment from 'moment';
 import moment_jalaali from 'moment-jalaali';
+import { AccessService } from "../../../service/service.access";
 
 //// props & state define ////////
 export interface IProps {
@@ -151,18 +152,22 @@ class PersonManageComponent extends BaseComponent<IProps, IState>{
           }
         },
       ],
-      actions: [
+      actions: this.checkAllAccess() ?  [
         {
+          access : (row : any) => {return this.checkDeleteToolAccess()},
           text: <i title={Localization.remove} className="fa fa-trash text-danger"></i>,
           ac_func: (row: any) => { this.onShowRemoveModal(row) },
           name: Localization.remove
         },
         {
+          access : (row : any) => {return this.checkUpdateToolAccess()},
           text: <i title={Localization.update} className="fa fa-pencil-square-o text-primary"></i>,
           ac_func: (row: any) => { this.updateRow(row) },
           name: Localization.update
         },
       ]
+      :
+      undefined
     },
     filter: {
       person: {
@@ -189,6 +194,27 @@ class PersonManageComponent extends BaseComponent<IProps, IState>{
     tableProcessLoader: false,
   }
 
+  checkAllAccess():boolean{
+    if(AccessService.checkOneOFAllAccess(['PERSON_DELETE_PREMIUM','PERSON_EDIT_PREMIUM'])){
+      return true;
+    }
+    return false;
+  }
+
+  checkDeleteToolAccess():boolean{
+    if(AccessService.checkAccess('PERSON_DELETE_PREMIUM')){
+      return true;
+    }
+    return false
+  }
+
+  checkUpdateToolAccess():boolean{
+    if(AccessService.checkAccess('PERSON_EDIT_PREMIUM')){
+      return true;
+    }
+    return false
+  }
+
   selectedPerson: IPerson | undefined;
   private _personService = new PersonService();
   // private _priceService = new PriceService();
@@ -199,6 +225,9 @@ class PersonManageComponent extends BaseComponent<IProps, IState>{
   // }
 
   updateRow(person_id: any) {
+    if(!AccessService.checkAccess('PERSON_EDIT_PREMIUM')){
+      return;
+    }
     this.props.history.push(`/person/${person_id.id}/edit`);
   }
 
@@ -227,6 +256,9 @@ class PersonManageComponent extends BaseComponent<IProps, IState>{
   /////// delete modal function define ////////
 
   onShowRemoveModal(person: IPerson) {
+    if(!AccessService.checkAccess('PERSON_DELETE_PREMIUM')){
+      return;
+    };
     this.selectedPerson = person;
     this.setState({ ...this.state, removeModalShow: true });
   }
@@ -238,6 +270,9 @@ class PersonManageComponent extends BaseComponent<IProps, IState>{
   }
 
   async onRemovePerson(person_id: string) {
+    if(!AccessService.checkAccess('PERSON_DELETE_PREMIUM')){
+      return;
+    };
     this.setState({ ...this.state, setRemoveLoader: true });
     let res = await this._personService.remove(person_id).catch(error => {
       this.handleError({ error: error.response });
@@ -431,6 +466,9 @@ class PersonManageComponent extends BaseComponent<IProps, IState>{
   //// navigation function //////
 
   gotoPersonCreate() {
+    if(!AccessService.checkAccess('PERSON_ADD_PREMIUM')){
+      return;
+    };
     this.props.history.push('/person/create');
   }
 
@@ -507,19 +545,25 @@ class PersonManageComponent extends BaseComponent<IProps, IState>{
     return (
       <>
         <div className="content">
-          <div className="row">
-            <div className="col-12">
-              <h2 className="text-bold text-dark pl-3">{Localization.person}</h2>
-              <BtnLoader
-                loading={false}
-                disabled={false}
-                btnClassName="btn btn-success shadow-default shadow-hover mb-4"
-                onClick={() => this.gotoPersonCreate()}
-              >
-                {Localization.new}
-              </BtnLoader>
-            </div>
-          </div>
+          {
+            AccessService.checkAccess('PERSON_ADD_PREMIUM')
+              ?
+              <div className="row">
+                <div className="col-12">
+                  <h2 className="text-bold text-dark pl-3">{Localization.person}</h2>
+                  <BtnLoader
+                    loading={false}
+                    disabled={false}
+                    btnClassName="btn btn-success shadow-default shadow-hover mb-4"
+                    onClick={() => this.gotoPersonCreate()}
+                  >
+                    {Localization.new}
+                  </BtnLoader>
+                </div>
+              </div>
+              :
+              undefined
+          }
           <div className="row">
             <div className="col-12">
               <div className="template-box mb-4">
